[feat] Move implementation to preprocess.py
